🎉 Flow Typing 🎉 (#3138)
* implemented Flow internally

* added external flow typing for exported interface

* added flowtyped lodash and fixed test coverage

* Removed some unused type definitions

+++ src/createReducer.js
@@ -1,3 +1,4 @@
+// @flow
 import {
   ARRAY_INSERT,
   ARRAY_MOVE,
@@ -36,6 +37,7 @@ import {
 } from './actionTypes'
 import createDeleteInWithCleanUp from './deleteInWithCleanUp'
 import plain from './structure/plain'
+import type { Action, Structure } from './types'
 
 const isReduxFormAction = action =>
   action &&
@@ -43,16 +45,18 @@ const isReduxFormAction = action =>
   action.type.length > prefix.length &&
   action.type.substring(0, prefix.length) === prefix
 
-const createReducer = structure => {
+function createReducer<M, L>(structure: Structure<M, L>) {
   const {
     deepEqual,
     empty,
+    forEach,
     getIn,
     setIn,
     deleteIn,
     fromJS,
     keys,
     size,
+    some,
     splice
   } = structure
   const deleteInWithCleanUp = createDeleteInWithCleanUp(structure)
@@ -115,7 +119,7 @@ const createReducer = structure => {
     return result
   }
 
-  const behaviors = {
+  const behaviors: { [string]: { (state: any, action: Action): M } } = {
     [ARRAY_INSERT](state, { meta: { field, index }, payload }) {
       return arraySplice(state, field, index, 0, payload)
     },
@@ -179,7 +183,7 @@ const createReducer = structure => {
       return arraySplice(state, field, 0, 0, payload)
     },
     [AUTOFILL](state, { meta: { field }, payload }) {
-      let result = state
+      let result: any = state
       result = deleteInWithCleanUp(result, `asyncErrors.${field}`)
       result = deleteInWithCleanUp(result, `submitErrors.${field}`)
       result = setIn(result, `fields.${field}.autofilled`, true)
@@ -294,7 +298,7 @@ const createReducer = structure => {
           // initialize action causes the field to become pristine. That effect
           // is what we want.
           //
-          keys(registeredFields).forEach(name => {
+          forEach(keys(registeredFields), name => {
             const previousInitialValue = getIn(previousInitialValues, name)
             const previousValue = getIn(previousValues, name)
 
@@ -311,7 +315,7 @@ const createReducer = structure => {
             }
           })
 
-          keys(newInitialValues).forEach(name => {
+          forEach(keys(newInitialValues), name => {
             const previousInitialValue = getIn(previousInitialValues, name)
             if (typeof previousInitialValue === 'undefined') {
               // Add new values at the root level.
@@ -462,7 +466,7 @@ const createReducer = structure => {
       fields.forEach(
         field => (result = deleteIn(result, `fields.${field}.touched`))
       )
-      const anyTouched = keys(getIn(result, 'registeredFields')).some(key =>
+      const anyTouched = some(keys(getIn(result, 'registeredFields')), key =>
         getIn(result, `fields.${key}.touched`)
       )
       result = anyTouched
@@ -502,17 +506,20 @@ const createReducer = structure => {
     }
   }
 
-  const reducer = (state = empty, action) => {
+  const reducer = (state: any = empty, action: Action) => {
     const behavior = behaviors[action.type]
     return behavior ? behavior(state, action) : state
   }
 
-  const byForm = reducer => (state = empty, action = {}) => {
+  const byForm = reducer => (
+    state: any = empty,
+    action: Action = { type: 'NONE' }
+  ) => {
     const form = action && action.meta && action.meta.form
     if (!form || !isReduxFormAction(action)) {
       return state
     }
-    if (action.type === DESTROY) {
+    if (action.type === DESTROY && action.meta && action.meta.form) {
       return action.meta.form.reduce(
         (result, form) => deleteInWithCleanUp(result, form),
         state
@@ -529,7 +536,7 @@ const createReducer = structure => {
   function decorate(target) {
     target.plugin = function plugin(reducers) {
       // use 'function' keyword to enable 'this'
-      return decorate((state = empty, action = {}) =>
+      return decorate((state: any = empty, action: Action = { type: 'NONE' }) =>
         Object.keys(reducers).reduce((accumulator, key) => {
           const previousState = getIn(accumulator, key)
           const nextState = reducers[key](
